tests(aggregate): use separate file
